+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/ui/EffectedVideoChatActivity.kt
@@ -8,7 +8,7 @@ import android.media.effect.EffectFactory
 import android.os.Build
 import android.os.Bundle
 import android.os.Handler
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.util.Log
 import android.view.View
 import android.view.WindowManager