+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/ui/MainActivity.kt
@@ -4,10 +4,10 @@ import android.Manifest
 import android.annotation.TargetApi
 import android.content.Intent
 import android.os.Bundle
-import android.support.design.widget.Snackbar
-import android.support.v7.app.AlertDialog
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.LinearLayoutManager
+import com.google.android.material.snackbar.Snackbar
+import androidx.appcompat.app.AlertDialog
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.LinearLayoutManager
 import android.util.Log
 import jp.shiguredo.sora.sample.R
 import jp.shiguredo.sora.sdk.util.SoraLogger